fixed: compile warnings in RenderSystemDX.cpp
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/trunk@24002 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/RenderSystemDX.cpp
@@ -26,7 +26,7 @@
 #include "RenderSystemDX.h"
 #include "utils/log.h"
 #include "utils/TimeUtils.h"
-#include "GUIWindowManager.h"/
+#include "GUIWindowManager.h"
 
 using namespace std;
 
@@ -103,7 +103,7 @@ bool CRenderSystemDX::ResetRenderSystem(int width, int height)
   m_nBackBufferHeight = height;
 
   CRect rc;
-  rc.SetRect(0, 0, width, height);
+  rc.SetRect(0, 0, (float)width, (float)height);
 
   SetViewPort(rc);
